WT-17699 Fix cache fill ratio statistic always reporting the < 25% bucket
The fill ratio was computed with integer division of two uint64_t values,
truncating to 0 for any cache fill below 100% and dumping every hard-eviction
request into the lt_25 bucket. Compute the ratio in floating point so the
25-50%, 50-75% and >= 75% buckets are populated correctly.

Add a regression test that places all eviction triggers above 50% of the cache,
so a correct build records only the upper buckets.

+++ src/evict/evict_thread.c
@@ -544,7 +544,7 @@ __evict_update_work(WT_SESSION_IMPL *session, bool *eviction_needed)
      * If application threads are blocked by data in cache, track the fill ratio.
      *
      */
-    uint64_t cache_fill_ratio = bytes_inuse / bytes_max;
+    double cache_fill_ratio = (double)bytes_inuse / (double)bytes_max;
     bool evict_is_hard = LF_ISSET(WT_EVICT_CACHE_HARD);
     if (evict_is_hard) {
         if (cache_fill_ratio < 0.25)
